+++ src/main/java/com/greeenai/greeenai/domain/auth/controller/LoginController.java
@@ -1,0 +1,26 @@
+package com.greeenai.greeenai.domain.auth.controller;
+
+import com.greeenai.greeenai.domain.auth.service.LoginService;
+import com.greeenai.greeenai.domain.member.dto.LoginRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/login")
+@RequiredArgsConstructor
+public class LoginController {
+
+    private final LoginService loginService;
+
+    @PostMapping
+    public ResponseEntity<Void> login(@Valid @RequestBody LoginRequest request, HttpServletResponse response) {
+        loginService.login(request, response);
+        return ResponseEntity.ok().build();
+    }
+}
