+++ src/main/java/com/greeenai/greeenai/domain/auth/repository/RefreshTokenRepository.java
@@ -1,0 +1,6 @@
+package com.greeenai.greeenai.domain.auth.repository;
+
+import com.greeenai.greeenai.domain.auth.domain.RefreshToken;
+import org.springframework.data.repository.CrudRepository;
+
+public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Long> {}
