+++ src/main/java/com/greeenai/greeenai/domain/member/dto/LoginRequest.java
@@ -1,8 +1,10 @@
 package com.greeenai.greeenai.domain.member.dto;
 
-import com.greeenai.greeenai.domain.member.domain.OAuthProvider;
-
+import com.greeenai.greeenai.domain.member.domain.OauthProvider;
 import jakarta.validation.constraints.NotBlank;
 
-public record LoginRequest(@NotBlank String name, @NotBlank String email, @NotBlank String oAuthId, @NotBlank OAuthProvider oAuthProvider) {
-}
+public record LoginRequest(
+        @NotBlank String name,
+        @NotBlank String email,
+        @NotBlank String oauthId,
+        @NotBlank OauthProvider oAuthProvider) {}
